requirements.txt
asyncmy==0.2.5
attr==0.3.1
cffi==1.15.0
click==8.1.2
colorama==0.4.4
cryptography==37.0.0
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
jwt==1.3.1
MarkupSafe==2.1.1
marshmallow==3.15.0
mysqlclient==2.1.0
packaging==21.3
pycparser==2.21
pyparsing==3.0.8
python-dotenv==0.20.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
